requirements.txt
python-dotenv>=1.0.0
streamlit>=1.34
langchain>=0.2.10
langchain-core>=0.2.10
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2
langchain-groq>=0.1.4
chromadb>=0.5.3
pypdf>=4.2.0
sentence-transformers>=2.7.0
